+++ shell-zsh/.config/zsh/aliases.zsh
@@ -6,3 +6,8 @@ alias archtime='archaeology-0-core --timeline'
 alias archwk='archaeology-0-core --this-week'
 alias archint='archaeology-0-core --by-intent'
 alias archsince='archaeology-0-core --since'
+
+# Workspace Intelligence aliases
+alias ws='workspace-view'
+alias wsa='workspace-view --active'
+alias wss='workspace-view --summary'